working

+++ src/include/messaging.h
@@ -34,7 +34,7 @@ extern QueueHandle_t Movement_Queue;
 
 extern volatile int32_t encoder1Count;
 extern volatile int32_t encoder2Count;
-extern volatile int switchArray[16];
+extern volatile byte switchArray[16];
 extern volatile int touchArray[12];
 
 extern void touch_deal_with_message(messageParts message);
